pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["srub <shani.rubovitch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
pylint = "^3.2.3"
ipython = "^8.26.0"
ipdb = "^0.13.13"
colorlog = "^6.8.2"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
